refactor to reader directory
