+++ csharp/test/Drivers/Databricks/Unit/DatabricksOperationStatusPollerTests.cs
@@ -18,8 +18,8 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Apache.Arrow.Adbc.Drivers.Databricks.CloudFetch;
 using Apache.Arrow.Adbc.Drivers.Databricks;
+using Apache.Arrow.Adbc.Drivers.Databricks.Reader;
 using Apache.Hive.Service.Rpc.Thrift;
 using Moq;
 using Xunit;
